Cmake build and warning fixing

+++ fcd/dumb_allocator.h
@@ -18,6 +18,7 @@
 #include <iterator>
 #include <list>
 #include <memory>
+#include <cstring>
 #include <type_traits>
 
 #include <iostream>
@@ -106,7 +107,7 @@ public:
 	{
 		if (end >= begin)
 		{
-			size_t size = end - begin;
+			size_t size = size_t(end - begin);
 			if (auto memory = allocateDynamic<char>(size + 1))
 			{
 				std::copy(begin, end, memory);
